[score] plda py

+++ Score/plda_score.sh
@@ -9,7 +9,7 @@
 if [ $# != 5 ]; then
   echo "Usage: plda.sh <lda-dim> <data-path> <train-feat-dir> <test-feat-dir> <trials>"
   echo "e.g.: plda.sh data/train exp/train exp/test data/test/trials"
-  echo "This script helps create srp to resample wav in wav.scp"
+  echo "This script helps plda scoring for ivectors"
   exit 1
 fi
 
@@ -26,20 +26,20 @@ logdir=$test_feat_dir/log
 
 test_score=$test_feat_dir/scores_$(date "+%Y-%m-%d-%H-%M-%S")
 
-if ! [ -f $train_feat_dir/utt2spk ];then
+if ! [ -s $train_feat_dir/utt2spk ];then
     echo "Creating utt2spk!"
     cat $train_feat_dir/xvectors.scp | awk '{print $1}' > $train_feat_dir/utt
     grep -f $train_feat_dir/utt $data_dir/utt2spk > $train_feat_dir/utt2spk
 fi
 
-if ! [ -f $train_feat_dir/spk2utt ];then
+if ! [ -s $train_feat_dir/spk2utt ];then
     Score/utt2spk_to_spk2utt.pl ${train_feat_dir}/utt2spk > $train_feat_dir/spk2utt
     num_spks=`wc -l $train_feat_dir/spk2utt | awk '{print $1}'`
     echo "There are ${num_spks} speakers in train set!"
 fi
 
 
-if ! [ -f $train_feat_dir/transform.mat ];then
+if ! [ -s $train_feat_dir/transform.mat ];then
   python Score/Plda/compute_lda.py --total-covariance-factor=0.0 \
     --lda-dim $lda_dim  \
     --spk2utt $data_dir/spk2utt \
@@ -50,7 +50,7 @@ if ! [ -f $train_feat_dir/transform.mat ];then
 fi
 
 
-if ! [ -f $train_feat_dir/plda ];then
+if ! [ -s $train_feat_dir/plda ];then
     python Score/Plda/compute_plda.py --spk2utt $train_feat_dir/spk2utt \
       --ivector-scp $train_feat_dir/xvectors.scp \
       --mean-vec $train_feat_dir/mean.vec \
@@ -58,25 +58,29 @@ if ! [ -f $train_feat_dir/plda ];then
       --plda-file $train_feat_dir/plda
 fi
 
-python Score/Plda/plda_scoring.py --normalize-length \
-  --train-vec-scp $test_feat_dir/xvectors.scp \
-  --test-vec-scp $test_feat_dir/xvectors.scp \
-  --trials $trials \
-  --plda-file $train_feat_dir/plda \
-  --transform-vec $train_feat_dir/transform.mat \
-  --score $test_score
+if ! [ -s $test_score ];then
+  python Score/Plda/plda_scoring.py --normalize-length \
+    --train-vec-scp $test_feat_dir/xvectors.scp \
+    --test-vec-scp $test_feat_dir/xvectors.scp \
+    --trials $trials \
+    --plda-file $train_feat_dir/plda \
+    --transform-vec $train_feat_dir/transform.mat \
+    --score $test_score
+fi
 
+if ! [ -s $test_score ];then
 
-eer=`compute-eer <(Score/prepare_for_eer.py $trials $test_score) 2> /dev/null`
-mindcf1=`Score/compute_min_dcf.py --p-target 0.01 $test_score $trials 2> /dev/null`
-mindcf2=`Score/compute_min_dcf.py --p-target 0.001 $test_score $trials 2> /dev/null`
+  eer=`compute-eer <(Score/prepare_for_eer.py $trials $test_score) 2> /dev/null`
+  mindcf1=`Score/compute_min_dcf.py --p-target 0.01 $test_score $trials 2> /dev/null`
+  mindcf2=`Score/compute_min_dcf.py --p-target 0.001 $test_score $trials 2> /dev/null`
 
-test_result=$test_feat_dir/result_plda_$(date "+%Y.%m.%d.%H-%M-%S")
+  test_result=$test_feat_dir/result_plda_$(date "+%Y.%m.%d.%H-%M-%S")
 
-echo "EER: $eer%" >> $test_result
-echo "minDCF(p-target=0.01): $mindcf1" >> $test_result
-echo "minDCF(p-target=0.001): $mindcf2" >> $test_result
+  echo "EER: $eer%" >> $test_result
+  echo "minDCF(p-target=0.01): $mindcf1" >> $test_result
+  echo "minDCF(p-target=0.001): $mindcf2" >> $test_result
 
-echo "EER: $eer%"
-echo "minDCF(p-target=0.01): $mindcf1"
-echo "minDCF(p-target=0.001): $mindcf2"
+  echo "EER: $eer%"
+  echo "minDCF(p-target=0.01): $mindcf1"
+  echo "minDCF(p-target=0.001): $mindcf2"
+fi
